major project/requirements.txt
# Malicious URL Detection Model Dependencies
# Python 3.10+

tensorflow>=2.13.0,<2.16.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
seaborn>=0.12.0,<1.0.0
matplotlib>=3.7.0,<4.0.0
pyyaml>=6.0,<7.0.0
joblib>=1.3.0,<2.0.0
pytest>=7.4.0,<8.0.0
